requirements.txt
absl-py==1.4.0
asttokens==2.2.1
attrs==23.1.0
backcall==0.2.0
cffi==1.15.1
cloudpickle==2.2.1
colorama==0.4.6
comm==0.1.4
contourpy==1.1.0
cycler==0.11.0
debugpy==1.6.7.post1
decorator==5.1.1
executing==1.2.0
flatbuffers==23.5.26
fonttools==4.42.1
ipykernel==6.25.1
ipython==8.15.0
jedi==0.19.0
joblib==1.3.2
jupyter_client==8.3.1
jupyter_core==5.3.1
kiwisolver==1.4.5
matplotlib==3.7.2
matplotlib-inline==0.1.6
mediapipe==0.10.3
nest-asyncio==1.5.7
numpy==1.25.2
opencv-contrib-python==4.8.0.76
opencv-python==4.8.0.76
packaging==23.1
pandas==2.1.0
parso==0.8.3
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.10.0
prompt-toolkit==3.0.39
protobuf==3.20.3
psutil==5.9.5
pure-eval==0.2.2
pycparser==2.21
Pygments==2.16.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
pywin32==306
pyzmq==25.1.1
scikit-learn==1.3.0
scipy==1.11.2
six==1.16.0
sounddevice==0.4.6
spyder-kernels==2.4.4
stack-data==0.6.2
threadpoolctl==3.2.0
tornado==6.3.3
traitlets==5.9.0
tzdata==2023.3
wcwidth==0.2.6
